always store all_atoms for init_walkers
A couple of changes which are not backwards incompatible.

First always store an `alt_reps/all_atoms` field for the initial
walkers. This is useful as it matches the topology and you would
probably always want this. This will also always store a
`_settings/alt_reps_idxs` record for the all atoms so other code can
work with it.

Second, adapt the `report` code so it doesn't save the all_atoms rep
unless specifically requested even though we are saving the all_atoms
atom idxs.