refactor work mapper runner interfaces
Start to reify some of the implicit interfaces the sim manager expects
from the runners/work mappers.

Secondly experimenting with having the runners be responsible for the
work mapper task generation. This makes sense since the runner knows
the specifics of how the tasks need to be parametrized. Seems to make sense.