new ala test system, cleanup
The Alanine dipeptide system builder from openmmtools was causing
issues with the simulation. It probably generated contexts or
something. Switched to using some XML serialized systems I had laying
around.

Added a lot of extra logging in the runner on the configuration and
simulation system state during the lifecycle for simpler debugging
during the simulations.