add clear types for resampl{ing,er} records
Convert to attrs classes, add validation, and distinguish between
decision records and resampler records.

Add an AttrsMappingMixin so that the Records can be treated like dicts
downstream.

Create an explicit storage protocol which defines the kinds of types
that can be generated for different things. In this case the records.